Speed up test execution
- Pass arguments directly to run_slide rather than starting a new shell
  process
  - Abstracts option generation by allowing variable parsing of the app
    arguments
- Delete debug test, as now all slide executions run in the same
  process, thereby making the interner non-deterministic because test
  order is also non-deterministic.

The benefit of this tests now run in 5-8 seconds rather than 30 on my
machine.
