--- slide/src/test/ui/emit/debug.slide
@@ -1,29 +1,0 @@
-!!!args
---parse-only
--o debug
-!!!args
-
-===in
-1 + 5 / 10 + 2
-===in
-
-~~~stdout
-Expr(
-    InternedExpr {
-        spur: Spur {
-            key: 7,
-        },
-        span: Span {
-            lo: 0,
-            hi: 14,
-        },
-    },
-)
-~~~stdout
-
-~~~stderr
-~~~stderr
-
-~~~exitcode
-0
-~~~exitcode
